Convert Tag to .tsx
