+++ frontend/src/components/General/Tag.test.tsx
@@ -3,11 +3,10 @@ import { render, fireEvent } from '@testing-library/react';
 
 import Tag from './Tag';
 
-it('logs the value and type of the tag with onClose()', (done) => {
+it('logs the value and type of the tag with onClose()', () => {
   // eslint-disable-next-line no-console
-  const onClose = () => done();
   const { getByRole } = render(
-    <Tag value="foo" type="bar" onClose={onClose}>
+    <Tag value="foo" type="bar" onClose={jest.fn()}>
       tag
     </Tag>
   );
